refactor(headless-docsite): move theme/ into the stories package
Owning tokens.css inside the stories package (where the headless story
authors edit theme variables) is cleaner than reaching from a sibling
package into the docsite app. The docsite now imports it from there.
